IT'S ALIVE...well not entry point

+++ Grok3d_Windows/src/Component/TransformComponent.cpp
@@ -42,7 +42,7 @@ int GRK_TransformComponent::GetChildIndex(GRK_TransformComponent* possibleChild)
 
 int GRK_TransformComponent::ChildCount()
 {
-    return this->m_children.size();
+    return static_cast<int>(this->m_children.size());
 }
 
 glm::vec3 GRK_TransformComponent::GetWorldPosition()
